fully integrated authentication -- validate method bug. Note: Server running at 3000 on this branch
