+++ lib/api/credit.js
@@ -12,6 +12,7 @@ exports.credits = {
   },
   create: function(request, reply) {
     models.Credits.create({
+      //api.auth.verify(request.token);
       createdAt: Date.now(),
       UserId: request.payload.userId,
       redeemed: false
