+++ lib/api/user.js
@@ -22,6 +22,7 @@ exports.users = {
   create: function(request, reply) {
     models.User.create(request.payload)
       .then(function (newUser) {
+        //TODO: this ORM is creating null users. Solution? Something to do with signup.html
         reply(newUser).code(200);
       })
       .catch(function (error){
