+++ lib/routes/redemption.js
@@ -5,6 +5,8 @@ module.exports = [
   {
     method: 'GET',
     path: '/api/redemptions',
+    handler: api.redemptions.all
+    ,config: {auth: true}
     handler: api.redemptions.all,
     config: {
       tags: ['api']
@@ -15,6 +17,7 @@ module.exports = [
     path: '/api/redemptions',
     handler: api.redemptions.create,
     config: {
+      auth: false,
       tags: ['api'],
       validate: {
         payload: {
