+++ lib/routes/user.js
@@ -23,6 +23,7 @@ module.exports = [
     path: '/api/users',
     handler: api.users.create,
     config: {
+      auth: false,
       tags: ['api'],
       validate: {
         payload: {
