+++ server.js
@@ -1,14 +1,16 @@
 'use strict';
 
-const Hapi        = require('hapi');
+const Hapi      = require('hapi');
+const env       = process.env.NODE_ENV || 'development';
+const config    = require(__dirname + '/config/config')[env];
 const Inert       = require('inert');
 const Vision      = require('vision');
 const HapiSwagger = require('hapi-swagger');
-const env         = process.env.NODE_ENV || 'development';
-const config      = require(__dirname + '/config/config')[env];
-const models      = require('./models');
+const models    = require('./models');
+const uuid      = require('uuid');
+const nJwt      = require('njwt');
 const Pack        = require('./package');
-const server      = new Hapi.Server();
+const server = new Hapi.Server();
 
 const options = {
   info: {
@@ -24,22 +26,15 @@ server.route(require('./lib/routes/credit'));
 server.route(require('./lib/routes/redemption'));
 server.route(require('./lib/routes/work-session'));
 server.route(require('./lib/routes/credit-transaction'));
-server.route(require('./lib/routes/login'));
-server.route(require('./lib/routes/logout'));
+server.route(require('./lib/routes/auth'));
 
-
-var users = { // collect users from db
-  1: {
-    id: 1,
-    name: 'Jen Jones'
-  }
-};
 
 //fill this out
 var validate = function (decoded, request, callback) {
-
-  // do your checks to see if the person is valid
-  if (!people[decoded.id]) {
+  var secretKey = uuid.v4();
+  var token = request.headers.tokenPassed;
+  var verifiedJwt = nJwt.verify(token,secretKey);
+  if (!verifiedJwt) {
     return callback(null, false);
   }
   else {
@@ -60,22 +55,6 @@ server.register(require('hapi-auth-jwt2'), function (err) {
       });
 
   server.auth.default('jwt');
-
-  server.route([
-    {
-      method: "GET", path: "/", config: { auth: false },
-      handler: function(request, reply) {
-        reply({text: 'Token not required'});
-      }
-    },
-    {
-      method: 'GET', path: '/restricted', config: { auth: 'jwt' },
-      handler: function(request, reply) {
-        reply({text: 'You used a Token!'})
-            .header("Authorization", request.headers.authorization);
-      }
-    }
-  ]);
 });
 
 server.register([ Inert, Vision, {
